Move emulator into its own screen
