+++ lib/ui/router/router.dart
@@ -1,10 +1,11 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:nesd/ui/emulator/main_screen.dart';
+import 'package:nesd/ui/emulator/emulator_screen.dart';
 import 'package:nesd/ui/emulator/rom_manager.dart';
 import 'package:nesd/ui/file_picker/file_picker_screen.dart';
 import 'package:nesd/ui/file_picker/file_system/filesystem_file.dart';
+import 'package:nesd/ui/main_menu/main_screen.dart';
 import 'package:nesd/ui/menu/menu_screen.dart';
 import 'package:nesd/ui/save_states/save_states_screen.dart';
 import 'package:nesd/ui/settings/controls/touch/touch_editor_screen.dart';
@@ -25,6 +26,13 @@ class Router extends RootStackRouter {
       transitionsBuilder: TransitionsBuilders.noTransition,
       duration: Duration.zero,
       reverseDuration: Duration.zero,
+    ),
+    CustomRoute(
+      page: EmulatorRoute.page,
+      path: '/emulator',
+      transitionsBuilder: TransitionsBuilders.fadeIn,
+      duration: const Duration(milliseconds: 200),
+      reverseDuration: const Duration(milliseconds: 200),
     ),
     AutoRoute(page: SettingsRoute.page, path: '/settings'),
     AutoRoute(page: TouchEditorRoute.page, path: '/touch_editor'),
